youtube_blog_generator/requirements.txt
streamlit>=1.32.0
yt-dlp>=2024.3.10
opencv-python-headless>=4.8.0
numpy>=1.24.0
torch>=2.2.0
transformers>=4.36.2
Pillow>=10.0.0
whisper>=1.1.10
anthropic>=0.18.0
reportlab>=4.0.8
python-dotenv>=1.0.0
scikit-learn>=1.3.0
soundfile>=0.12.1 